ymir-sdl3/App: Move screenshot stuff to ScreenshotService
